requirements.txt
python-dotenv==1.0.0
discord.py-self==2.1.0
